Moved Iterator and String steppers to their own files.
